comments - remote or rename flow subdir
