--- src/promptflow-evals/promptflow/evals/evaluators/_f1_score/flow/__init__.py
@@ -1,12 +1,0 @@
-# ---------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# ---------------------------------------------------------
-
-# Relative imports don't work for loaded evaluators, so we need absolute imports to be possible.
-from .f1_score import compute_f1_score
-from .validate_inputs import validate_inputs
-
-__all__ = [
-    "compute_f1_score",
-    "validate_inputs",
-]
