--- src/promptflow-evals/promptflow/evals/evaluators/_f1_score/flow/data.jsonl
@@ -1,1 +1,0 @@
-{"groundtruth": "App", "prediction": "App"}
