--- src/promptflow-evals/promptflow/evals/evaluators/_f1_score/flow/requirements.txt
@@ -1,2 +1,0 @@
-promptflow
-promptflow-tools
